Disable send's directory redirect feature
It doesn't work well enough with prefixed routes due to Medley's routing. It's easier just to disable it entirely and consistently return a 404 for requests with a URL that doesn't end in a '/'.

Future Note:
The reason it doesn't work well is that the Medley app may or may not be using the 'strictRouting' option. To properly configure the serve-static plugin, the plugin would need to know the value of this option (so the app would need to expose it). Then serve-static would need to define different routes depending on the value of the option and pass the right pathname to the send module. It is doable, but would require several if-else statements and a bunch more tests.